+++ client/CHANGELOG.md
@@ -27,6 +27,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 - Toggle to set the visibility (public or private) of a scenario. [LANDGRIF-1009](https://vizzuality.atlassian.net/browse/LANDGRIF-1009)
 
+### Fixed
+
+- Tests were failing due to a change in the dev environment
+
 ## [v0.3.0]
 
 ### Added