+++ client/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Fixed
 
+- Fixed CHILD select strategy not working with empty children list [LANDGRIF-1079](https://vizzuality.atlassian.net/browse/LANDGRIF-1079)
 - Update rows expanded state on analysis table [LANDGRIF-967](https://vizzuality.atlassian.net/browse/LANDGRIF-967)
 - Horizontal scroll on the legend [LANDGRIF-1082](https://vizzuality.atlassian.net/browse/LANDGRIF-1082)
 - Format of numbers of the contextual layers in the legend [LANDGRIF-1090](https://vizzuality.atlassian.net/browse/LANDGRIF-1090)
@@ -86,6 +87,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - Added interaction in the charts clicking on the legend [LANDGRIF-772](https://vizzuality.atlassian.net/browse/LANDGRIF-772)
 
 ### Changed
+
 - Removed minus sign (`-`) from absolute difference in comparison charts. [LANDGRIF-1046](https://vizzuality.atlassian.net/browse/LANDGRIF-1046)
 - Bigger table page size [LANDGRIF-922](https://vizzuality.atlassian.net/browse/LANDGRIF-922)
 - `ACTUAL_DATA` was removed and a null scenario id is used in it's place